src/server/requirements.txt
appdirs==1.4.3
click==6.7
Flask==0.12.1
Flask-Cors==3.0.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
nltk==3.2.2
numpy==1.12.1
oauthlib==2.0.2
packaging==16.8
pandas==0.19.2
pandas-datareader==0.3.0.post0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
requests-file==1.4.1
requests-ftp==0.3.1
requests-oauthlib==0.8.0
scikit-learn==0.18.1
scipy==0.18.1
simplejson==3.10.0
six==1.10.0
tweepy==3.5.0
Werkzeug==0.12.1
yahoo-finance==1.4.0
